cleaning some duplicated macros
